refactor(core): extract auth-failure event helpers from collector
Move the six pure auth-failure helpers and their two constants out of
collector.py into a new orchestration/auth_failure.py. collector.py drops
from 772 to 664 lines; the AuthFailed event builder and its PII scrubbers
(_strip_url_query, _scrub_password) now live together with their own
responsibility. Pure move, no behavior change.
